dataLater now rendered by Layout component
