+++ src/components/index.ts
@@ -14,3 +14,4 @@ export { TopHeader } from "./topHeader/topHeader"
 export { TriangleLinks } from "./triangleLinks/triangleLinks"
 export { PreviousPageLink } from "./previousPageLink/previousPageLink"
 export { NextPageLink } from "./nextPageLink/nextPageLink"
+export { DataLayer } from "./dataLayer/dataLayer"
