+++ src/components/layout/index.tsx
@@ -1,57 +1,29 @@
 import React, { ReactNode } from "react"
-import { graphql, useStaticQuery } from "gatsby"
 
-import { NavBar, TriangleLinks } from "../index"
+import { DataLayer } from "../index"
 import "./index.scss"
+
+interface FrontmatterData {
+  title: string
+  jobTitle: string
+  greeting: string
+}
 
 interface Props {
   children: ReactNode
   pathName: string
+  listLinks: Array<string>
+  frontmatterData: Array<FrontmatterData>
 }
 
 const Layout = (props: Props) => {
   const { children, pathName } = props
 
-  const layoutData = useStaticQuery(graphql`
-    {
-      allMarkdownRemark(sort: { fields: frontmatter___chapter }) {
-        edges {
-          node {
-            fields {
-              slug
-            }
-            frontmatter {
-              title
-              jobTitle
-              greeting
-            }
-          }
-        }
-      }
-    }
-  `)
-
-  const listLinks = layoutData.allMarkdownRemark.edges.map(
-    ({ node }: { node: any }, index: number) => node.fields.slug
-  )
-
-  const frontmatterData = layoutData.allMarkdownRemark.edges.map(
-    ({ node }: { node: any }, index: number) => node.frontmatter
-  )
-
   return (
-
     <div>
-        <NavBar
-          pathName={pathName}
-          listLinks={listLinks}
-          frontmatterData={frontmatterData}
-        />
-      <div className="d-flex justify-content-center">
-        <TriangleLinks pathName={pathName} listLinks={listLinks}/>
-      </div>
-        <main>{children}</main>
-      </div>
+      <DataLayer pathName={pathName}/>
+      <main>{children}</main>
+    </div>
   )
 }
